--- libs/backend/abstract/ports/README.md
@@ -1,3 +1,0 @@
-# backend-abstract-ports
-
-This library was generated with [Nx](https://nx.dev).
