--- libs/backend/abstract/ports/project.json
@@ -1,24 +1,0 @@
-{
-  "name": "backend-abstract-ports",
-  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/backend/abstract/ports/src",
-  "projectType": "library",
-  "targets": {
-    "lint": {
-      "executor": "@nx/eslint:lint",
-      "configurations": {
-        "ci": {
-          "format": "junit",
-          "outputFile": "tmp/reports/lint/backend-abstract-ports.xml",
-          "quiet": true
-        }
-      },
-      "options": {
-        "lintFilePatterns": [
-          "libs/backend/abstract/ports/**/*.{ts,tsx,js,jsx,graphql}"
-        ]
-      }
-    }
-  },
-  "tags": ["projectType:library", "type:abstract", "scope:backend"]
-}
