--- libs/backend/abstract/ports/src/index.ts
@@ -1,1 +1,0 @@
-export * from './type-transformer'
