+++ libs/backend/abstract/types/src/index.ts
@@ -3,3 +3,4 @@ export * from './model.interface'
 export * from './neo4j.repo.interface'
 export * from './repository.interface'
 export * from './server.interface'
+export * from './type-transformer.interface'
