+++ libs/backend/abstract/types/src/server.interface.ts
@@ -1,9 +1,15 @@
 import type { Auth0IdToken } from '@codelab/shared/abstract/core'
+import type { IncomingMessage } from 'http'
 import type { NextApiRequest as OriginalNextApiRequest } from 'next'
 
 export interface GraphQLRequestContext {
   req: NextApiRequest
   user?: Auth0IdToken
+}
+
+export interface GqlContext {
+  req: IncomingMessage
+  res: Response
 }
 
 interface NextApiRequest extends OriginalNextApiRequest {
