+++ libs/backend/infra/adapter/graphql/src/index.ts
@@ -2,4 +2,3 @@ export * from './graphql.client'
 export * from './graphql.module'
 export * from './graphql.service'
 export * from './middleware'
-export * from './tests/setup'
