--- libs/backend/infra/adapter/graphql/src/middleware/context.ts
@@ -1,6 +1,0 @@
-import type { IncomingMessage } from 'http'
-
-export interface GqlContext {
-  req: IncomingMessage
-  res: Response
-}
