+++ libs/backend/infra/adapter/graphql/src/middleware/index.ts
@@ -1,4 +1,3 @@
 export * from './auth.middleware'
-export * from './context'
 export * from './cors.middleware'
 export * from './proxy.middleware'
