+++ libs/backend/infra/adapter/neo4j/src/index.ts
@@ -5,3 +5,4 @@ export * from './neo4j.config'
 export * from './resolver'
 export * from './schema'
 export * from './selectionSet'
+export * from './test/setup'
