+++ libs/backend/infra/adapter/neo4j/src/resolver/ogm-resolver/component/component.resolver.i.spec.ts
@@ -1,22 +1,18 @@
 import { AtomType } from '@codelab/backend/abstract/codegen'
-import { OgmService } from '@codelab/backend/infra/adapter/neo4j'
-import { IPageKind, IPageKindName } from '@codelab/shared/abstract/core'
-import { ROOT_ELEMENT_NAME } from '@codelab/shared/config'
 import {
-  AppProperties,
   connectNodeId,
   connectOwner,
   ElementProperties,
-  PageProperties,
 } from '@codelab/shared/domain'
 import type { INestApplication } from '@nestjs/common'
 import { print } from 'graphql'
 import request from 'supertest'
 import { v4 } from 'uuid'
-import { PageResolverPages } from './page.spec.graphql.gen'
-import { setupTestingContext } from './setup'
+import { OgmService } from '../../../infra'
+import { setupTestingContext } from '../../../test/setup'
+import { ComponentResolverComponents } from './component.spec.graphql.gen'
 
-describe('ComponentResolvers', () => {
+describe('PageResolvers', () => {
   let app: INestApplication
   let ogmService: OgmService
   const context = setupTestingContext()
@@ -37,7 +33,7 @@ describe('ComponentResolvers', () => {
     await ctx.afterAll()
   })
 
-  it('should fetch a page with field resolvers - name, slug, elements', async () => {
+  it('should fetch a component with elements resolver', async () => {
     const owner = (
       await ogmService.User.create({
         input: [
@@ -51,19 +47,7 @@ describe('ComponentResolvers', () => {
       })
     ).users[0]!
 
-    const testApp = (
-      await ogmService.App.create({
-        input: [
-          {
-            compositeKey: AppProperties.appCompositeKey('My App', owner),
-            id: v4(),
-            owner: connectOwner(owner),
-          },
-        ],
-      })
-    ).apps[0]!
-
-    const testPageRef = { id: v4() }
+    const componentRef = { id: v4() }
 
     const atomApi = (
       await ogmService.InterfaceType.create({
@@ -107,8 +91,8 @@ describe('ComponentResolvers', () => {
         input: [
           {
             compositeKey: ElementProperties.elementCompositeKey(
-              ROOT_ELEMENT_NAME,
-              testPageRef,
+              'Component Root',
+              componentRef,
             ),
             id: v4(),
             props: connectNodeId(props.id),
@@ -137,7 +121,7 @@ describe('ComponentResolvers', () => {
           {
             compositeKey: ElementProperties.elementCompositeKey(
               'Child Element',
-              testPageRef,
+              componentRef,
             ),
             id: v4(),
             parentElement: connectNodeId(rootElement.id),
@@ -174,33 +158,53 @@ describe('ComponentResolvers', () => {
       })
     ).stores[0]!
 
-    const testPage = (
-      await ogmService.Page.create({
+    const componentApi = (
+      await ogmService.InterfaceType.create({
         input: [
           {
-            app: connectNodeId(testApp.id),
-            compositeKey: PageProperties.pageCompositeKey(
-              IPageKindName.Provider,
-              testApp,
-            ),
             id: v4(),
-            kind: IPageKind.Provider,
-            rootElement: connectNodeId(rootElement.id),
-            store: connectNodeId(store.id),
-            urlPattern: IPageKindName.Provider,
+            name: 'Component Api',
+            owner: connectOwner(owner),
           },
         ],
       })
-    ).pages[0]!
+    ).interfaceTypes[0]!
+
+    const componentProps = (
+      await ogmService.Prop.create({
+        input: [
+          {
+            data: '{}',
+            id: v4(),
+          },
+        ],
+      })
+    ).props[0]!
+
+    const component = (
+      await ogmService.Component.create({
+        input: [
+          {
+            api: connectNodeId(componentApi.id),
+            id: v4(),
+            name: 'Component',
+            owner: connectOwner(owner),
+            props: connectNodeId(componentProps.id),
+            rootElement: connectNodeId(rootElement.id),
+            store: connectNodeId(store.id),
+          },
+        ],
+      })
+    ).components[0]!
 
     await request(app.getHttpServer())
       .post('/graphql')
       .send({
-        query: print(PageResolverPages),
+        query: print(ComponentResolverComponents),
       })
       .expect(200)
       .expect((res) => {
-        expect(res.body.data.pages).toEqual([
+        expect(res.body.data.components).toEqual([
           {
             elements: [
               {
@@ -210,12 +214,11 @@ describe('ComponentResolvers', () => {
                 id: childElement.id,
               },
             ],
-            id: testPage.id,
-            name: IPageKindName.Provider,
+            id: component.id,
+            name: component.name,
             rootElement: {
               id: rootElement.id,
             },
-            slug: IPageKindName.Provider,
           },
         ])
       })
