+++ libs/backend/infra/adapter/neo4j/src/resolver/ogm-resolver/component/component.spec.graphql
@@ -1,11 +1,11 @@
 query componentResolverComponents {
   components {
+    elements {
+      id
+    }
     id
     name
     rootElement {
-      id
-    }
-    elements {
       id
     }
   }
