+++ libs/backend/infra/adapter/neo4j/src/resolver/ogm-resolver/page/page.spec.graphql
@@ -1,13 +1,13 @@
 query pageResolverPages {
   pages {
-    id
-    name
-    slug
-    rootElement {
-      id
-    }
     elements {
       id
     }
+    id
+    name
+    rootElement {
+      id
+    }
+    slug
   }
 }
