+++ tsconfig.base.json
@@ -30,9 +30,6 @@
       "@codelab/backend/abstract/core": [
         "libs/backend/abstract/core/src/index.ts"
       ],
-      "@codelab/backend/abstract/ports": [
-        "libs/backend/abstract/ports/src/index.ts"
-      ],
       "@codelab/backend/abstract/types": [
         "libs/backend/abstract/types/src/index.ts"
       ],
